Reorganize files within providers/DateStore
