+++ src/components/Tasks/TasksMonthSelector.tsx
@@ -6,7 +6,7 @@ import { DateFormat } from '@utils/i18n/date/patterns'
 import { BookkeepingPeriodStatus, useBookkeepingPeriods } from '@hooks/api/businesses/[business-id]/bookkeeping/periods/useBookkeepingPeriods'
 import { useEmitLayerEvent } from '@hooks/useEmitLayerEvent'
 import { useIntlFormatter } from '@hooks/utils/i18n/useIntlFormatter'
-import { useGlobalDate, useGlobalDatePeriodAlignedActions } from '@providers/DateStore/GlobalDateStore/GlobalDateStoreProvider'
+import { useGlobalDate, useGlobalDatePeriodAlignedActions } from '@providers/DateStore/GlobalDateStoreProvider'
 import { LayerEventComponent, LayerEventType } from '@providers/LayerProvider/layerEvents'
 import { TaskMonthTile } from '@components/Tasks/TaskMonthTile'
 import { type MonthData } from '@components/Tasks/types'
