+++ src/hooks/features/profitAndLoss/useProfitAndLossComparison.tsx
@@ -14,7 +14,7 @@ import { range } from '@utils/array/range'
 import { toLocalDateString } from '@utils/time/timeUtils'
 import { useProfitAndLossComparisonReport } from '@hooks/api/businesses/[business-id]/reports/profit-and-loss-comparison/useProfitAndLossComparisonReport'
 import { useAuth } from '@hooks/utils/auth/useAuth'
-import { useGlobalDateRange } from '@providers/DateStore/GlobalDateStore/GlobalDateStoreProvider'
+import { useGlobalDateRange } from '@providers/DateStore/GlobalDateStoreProvider'
 import { useEnvironment } from '@providers/Environment/EnvironmentInputProvider'
 import { useLayerContext } from '@contexts/LayerContext/LayerContext'
 import { ProfitAndLossContext } from '@contexts/ProfitAndLossContext/ProfitAndLossContext'
