+++ src/providers/LayerProvider/LayerProvider.tsx
@@ -8,7 +8,7 @@ import { DEFAULT_SWR_CONFIG } from '@utils/swr/defaultSWRConfig'
 import { localeKeyMiddleware } from '@utils/swr/localeKeyMiddleware'
 import { AuthInputProvider } from '@providers/AuthInputProvider'
 import { BusinessProvider } from '@providers/BusinessProvider/BusinessProvider'
-import { GlobalDateStoreProvider } from '@providers/DateStore/GlobalDateStore/GlobalDateStoreProvider'
+import { GlobalDateStoreProvider } from '@providers/DateStore/GlobalDateStoreProvider'
 import type { Environment, EnvironmentConfigOverride } from '@providers/Environment/environmentConfigs'
 import { EnvironmentInputProvider } from '@providers/Environment/EnvironmentInputProvider'
 import { LayerI18nProvider } from '@providers/I18nProvider/LayerI18nProvider'
