+++ src/views/ProjectProfitability/ProjectProfitability.tsx
@@ -6,7 +6,7 @@ import { DisplayState } from '@internal-types/bankTransactions'
 import { type MoneyFormat } from '@internal-types/general'
 import { translationKey } from '@utils/i18n/translationKey'
 import { type PnlTagFilter } from '@hooks/features/profitAndLoss/useProfitAndLoss'
-import type { DateSelectionMode } from '@providers/DateStore/GlobalDateStore/GlobalDateStoreProvider'
+import type { DateSelectionMode } from '@providers/DateStore/GlobalDateStoreProvider'
 import { HStack } from '@ui/Stack/Stack'
 import { Toggle } from '@ui/Toggle/Toggle'
 import { BankTransactions } from '@components/BankTransactions/BankTransactions'
